structure ColorCar,EngineCar,ModelCar,PowerCar,TypeCarBody - done

+++ src/main/java/autodealer/com/logic/dto/EngineCarDTO.java
@@ -7,6 +7,6 @@ import lombok.*;
 @AllArgsConstructor
 @RequiredArgsConstructor
 public class EngineCarDTO {
-    private int ID;
+    private Long ID;
     private String type_engine;
 }
